requirements.txt
fastapi
uvicorn
python-multipart
docling
psycopg2-binary
sqlalchemy
alembic
python-dotenv
pydantic
aiofiles
jinja2
# python-magic  # Optional - may not work on Windows